# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UMS API DOCUMENTATION",
        "description": "A simple laravel user management system for users",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "description": "This can only be done by an guest.",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessible Entity"
                    }
                }
            }
        },
        "login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Login in a User",
                "operationId": "userLogin",
                "requestBody": {
                    "description": "Updated user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unathorize Access"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/roles": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get all roles",
                "description": "Returns list of roles",
                "operationId": "getAllRoles",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Create role",
                "description": "This can only be done by an admin.",
                "operationId": "createRole",
                "requestBody": {
                    "description": "Create role object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRole"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessible Entity"
                    }
                }
            }
        },
        "/roles/{role_id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Get role by role id",
                "operationId": "getRoleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Role"
                ],
                "summary": "Delete role",
                "description": "This can only be done by an admin user.",
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Role not found"
                    },
                    "200": {
                        "description": "Successfull operation"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Role"
                ],
                "summary": "Updated role",
                "description": "This can only be done by an admin user.",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated role object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRole"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "bad request"
                    },
                    "404": {
                        "description": "Role not found"
                    },
                    "422": {
                        "description": "Unprocessible Entity"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{uuid}": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Updated user",
                "description": "This can only be done by an auth users.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "bad request"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Unprocessible Entity"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateRole": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "description": "role name",
                        "type": "string"
                    },
                    "description": {
                        "description": "role description",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Role": {
                "properties": {
                    "id": {
                        "description": "role id",
                        "type": "integer"
                    },
                    "title": {
                        "description": "role name",
                        "type": "string"
                    },
                    "description": {
                        "description": "role description",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "role description",
                        "type": "string"
                    },
                    "updated_at": {
                        "description": "role description",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateRole": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "description": "role id",
                        "type": "integer"
                    },
                    "title": {
                        "description": "role name",
                        "type": "string"
                    },
                    "description": {
                        "description": "role description",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "middlename": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "integer"
                    },
                    "profile_url": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateUser": {
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "middlename": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "integer"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUser": {
                "required": [
                    "uuid"
                ],
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "integer"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserLogin": {
                "required": [
                    "email",
                    "password",
                    "role_id"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}